Move traits to their own file
